Merge Intervals 02: Insert Intervals
Given a sorted list of nonoverlapping intervals and a new interval, your task is to insert the new interval into the correct position while ensuring that the resulting list of intervals remains sorted and nonoverlapping.
Each interval is a pair of nonnegative numbers, the first being the start time and the second being the end time of the interval.