Merge Intervals 04: Employee Free Time
You’re given a list containing the schedules of multiple employees.
Each person’s schedule is a list of non-overlapping intervals in sorted order.
An interval is specified with the start and end time, both being positive integers.
Your task is to find the list of intervals representing the free time for all the employees.